% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parseExample}
\alias{parseExample}
\title{Floating Point Parsing Example}
\usage{
parseExample(input = "3.1416 xyz ", verbose = TRUE)
}
\arguments{
\item{input}{A character variable with text to parse including a simple default}

\item{verbose}{A boolean variable to show or suppress progress, defaults to true}
}
\value{
A floating point scalar is returned on success; in case of parsing failure
the function exists via \code{stop()}.
}
\description{
This example is adapted from the example of the upstream README.md file, and generalized
to be called from R with variable input.
}
\examples{
parseExample()
}
